update tests to original location(s)
