+++ dlp/noxfile.py
@@ -3,7 +3,6 @@ import os
 
 import nox
 
-
 BLACK_VERSION = "black==19.3b0"
 BLACK_PATHS = ["tests", "noxfile.py"]
 
@@ -15,7 +14,7 @@ if os.path.exists("samples"):
 def setup(session):
     # Same as pip install -r -r requirements.txt.
     session.install("-r", "requirements.txt")
-    session.install("mock", "pytest")
+    session.install("mock", "pytest", "decorator")
 
     if not os.environ.get("GOOGLE_APPLICATION_CREDENTIALS", ""):
         session.skip("Credentials must be set via environment variable")
@@ -28,7 +27,7 @@ def tests_should_pass(session):
         "SHOULD_PASS_VPCSC": "True",
         "GCLOUD_PROJECT": "python-docs-samples-tests",
     }
-    tests_path = "tests"
+    tests_path = "."
     if not os.path.exists(tests_path):
         session.skip("tests were not found")
     session.run("pytest", "--quiet", tests_path, env=env)
@@ -41,7 +40,7 @@ def vpcsc_inspect_content_test(session):
         "SHOULD_PASS_VPCSC": "False",
         "GCLOUD_PROJECT": "vpcsc-dlp-1569864437-dut-0",
     }
-    tests_path = "tests/inspect_content_test.py"
+    tests_path = "./inspect_content_test.py"
     if not os.path.exists(tests_path):
         session.skip("tests were not found")
     session.run("pytest", "--quiet", tests_path, env=env)
@@ -54,7 +53,7 @@ def vpcsc_triggers_test(session):
         "SHOULD_PASS_VPCSC": "False",
         "GCLOUD_PROJECT": "vpcsc-dlp-1569864437-dut-0",
     }
-    tests_path = "tests/triggers_test.py"
+    tests_path = "./triggers_test.py"
     if not os.path.exists(tests_path):
         session.skip("tests were not found")
     session.run("pytest", "--quiet", tests_path, env=env)
