+++ dlp/quickstart_test.py
@@ -12,11 +12,25 @@
 # See the License for the specific language governing permissions and
 # limitations under the License.
 
-import dlp.metadata as metadata
+import os
+
+import google.cloud.dlp
+import mock
+
+import quickstart
 
 
-def test_fetch_info_types(capsys):
-    metadata.list_info_types()
+GCLOUD_PROJECT = os.getenv('GCLOUD_PROJECT')
+
+
+def test_quickstart(capsys):
+    # Mock out project_path to use the test runner's project ID.
+    with mock.patch.object(
+            google.cloud.dlp.DlpServiceClient,
+            'project_path',
+            return_value='projects/{}'.format(GCLOUD_PROJECT)):
+        quickstart.quickstart(GCLOUD_PROJECT)
 
     out, _ = capsys.readouterr()
-    assert 'EMAIL_ADDRESS' in out
+    assert 'FIRST_NAME' in out
+    assert 'LAST_NAME' in out
